KNOWLEDGE完成版

+++ wp-content/themes/WebPTemplatebyMarcat/include_files/single/main/cntMain.php
@@ -4,7 +4,7 @@
         <?php $nowCats = get_the_category();  ?>
         <a class="cl_93A5B1 bmtToArchiveCat" href="<?php echo get_category_link($nowCats[0]->cat_ID); ?>">一覧へ戻る</a>
         <div class="bookingPostKnowledgeLoop btmBookingClipsBookings">
-            <span class="iconCountBookingsLoop jsiconCountBookingsLoop"><?php echo SCF::get('countPosts'); ?></span>
+            <span class="iconCountBookingsLoop jsiconCountBookingsLoop"><?php echo (int)SCF::get('countPosts'); ?></span>
         </div>
         <div class="clipsPostKnowledgeLoop btmBookingClipsClips">
             <span class="iconClipsLoop jsiconClipsLoop" data-postID="<?php echo $post->ID; ?>">クリップする</span>
